requirements.txt
pygtk==2.24.0
pygobject==2.28.3
pycairo==1.8.10
click==7.1.2
PyYAML==5.4.1
